Refactor import components
