+++ src/components/index.js
@@ -1,0 +1,8 @@
+export { default as ProgressBar } from './ProgressBar'
+export { default as PieChart } from './PieChart'
+export { default as Navbar } from './Navbar'
+export { default as MathElement } from './MathElement'
+export { default as LoadingSpinner } from './LoadingSpinner'
+export { default as Kitem } from './Kitem'
+export { default as Jumbotron } from './Jumbotron'
+export { default as BarChart } from './BarChart'
