+++ src/containers/Schools/index.jsx
@@ -2,8 +2,10 @@ import React from 'react'
 import { browserHistory } from 'react-router'
 import { Row, Col } from 'react-flexbox-grid'
 import { useSchools, useStats } from '../../hooks'
-import Jumbotron from '../../components/Jumbotron'
-import Kitem from '../../components/Kitem'
+import {
+  Kitem,
+  Jumbotron,
+} from '../../components'
 import { header, name } from './methods'
 import styles from './Schools.css'
